+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-## [1.7.0] - 2025-08-XX
+## [1.7.0] - 2025-08-04
 
 - "translate" function moved from index.ts to translate.ts
 - Added variable consistency validation for {{ variableName }} patterns in translations